+++ test/MultiplexingSocketServerAppender-tests.js
@@ -11,30 +11,34 @@ describe("MultiplexingSocketAppender-tests", () => {
     it("configure", (done) => {
         Promise.resolve().then(async () => {
             const dir = await fse.mkdtemp("test")
-            const filename = path.join(dir, "test.log"), pattern = "yyyy-MM-dd_hh_mm_ss", listener = (event) => {
-                if (event.data[0] == "close") {
-                    log4js.shutdown(async (err) => {
-                        await fse.remove(dir)
-                        if (err) {
-                            console.error(err)
-                            return
-                        }
-                        done()
-                    })
+            const filename = path.join(dir, "test.log"),
+                pattern = "yyyy-MM-dd_hh_mm_ss",
+                listener = (event) => {
+                    if (event.level.levelStr == "EVENT" && event.data[0] == "close") {
+                        writer.destroy()
+                        log4js.shutdown(async (err) => {
+                            await fse.remove(dir)
+                            if (err) {
+                                console.error(err)
+                                return
+                            }
+                            done()
+                        })
+                        return false
+                    }
                 }
-            }
             log4js.configure({
                 appenders: {
-                    file: MultiplexingFileAppender.createConfig(filename, pattern, listener),
-                    server: MultiplexingSocketServerAppender.createConfig(1234)
+                    file: MultiplexingFileAppender.createConfig(filename, pattern),
+                    server: MultiplexingSocketServerAppender.createConfig(1234, listener)
                 },
                 categories: {
                     default: { appenders: ["file"], level: "all" }
                 }
             })
             const writer = new AutoReconnectMessageSocketWriter("127.0.0.1", 1234)
-            writer.send(new LoggingEvent("test", "INFO", ["hello", "world", "!"]))
-            writer.send(new LoggingEvent("test", "EVENT", ["close"]))
+            writer.send(new LoggingEvent("test", { levelStr: "INFO" }, ["hello", "world", "!"]))
+            writer.send(new LoggingEvent("test", { levelStr: "EVENT" }, ["close"]))
         })
     })
 })
